app/requirements.txt
psycopg2-binary
chromadb
streamlit
mistralai
postgres
vanna